requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.2.0
pandas==2.1.4
numpy==1.26.3
aiohttp==3.9.1
openai==1.12.0
pandasai==2.0.0
